geminiworkdeneme - Kopya/requirements.txt
flask==2.3.3
google-generativeai==0.3.1
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
gunicorn==21.2.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1